Move CKPersistable conformance to separate file
Extracted CloudKit persistence logic and related MapError enum from Matrix.swift into a new Matrix+CKPersistable.swift file. This improves code organization by separating persistence concerns from the core Matrix implementation.
